Gigantic refactor
