+++ CMakeLists.txt
@@ -27,7 +27,7 @@ file(GLOB_RECURSE TOTOGL_SOURCE_FILES CONFIGURE_DEPENDS ${TOTOGL_SOURCE_DIR}/*.c
 file(GLOB_RECURSE TOTOGL_TEST_INCLUDE_FILES CONFIGURE_DEPENDS ${TOTOGL_TEST_INCLUDE_DIR}/*.hpp)
 file(GLOB_RECURSE TOTOGL_TEST_SOURCE_FILES CONFIGURE_DEPENDS ${TOTOGL_TEST_SOURCE_DIR}/*.cpp)
 
-add_library(${PROJECT_NAME} SHARED ${TOTOGL_INCLUDE_FILES})
+add_library(${PROJECT_NAME} SHARED ${TOTOGL_SOURCE_FILES})
 target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_23)
 
 set_target_properties(${PROJECT_NAME} PROPERTIES LIBRARY_OUTPUT_DIRECTORY ${TOTOGL_BINARY_DIR})
