+++ include/TotoGL/RenderObject/Light.hpp
@@ -15,20 +15,11 @@ enum class LightType {
 class Light {
 public:
     Light() = delete;
-    Light(glm::vec3 color, float strength, LightType type)
-        : _type(type)
-        , _color(color)
-        , _strength(strength) { }
+    Light(glm::vec3 color, float strength, LightType type);
     ~Light() = default;
 
-    Light& translate(const glm::vec3& translation) {
-        _transform.translate(translation);
-        return *this;
-    }
-    Light& rotate(const float& angle, const glm::vec3& axis) {
-        _transform.rotate(angle, axis);
-        return *this;
-    }
+    Light& translate(const glm::vec3& translation);
+    Light& rotate(const float& angle, const glm::vec3& axis);
 
     LightType& type() { return _type; }
     glm::vec3& color() { return _color; }
@@ -36,11 +27,7 @@ public:
 
     Transform& transformation() { return _transform; }
     glm::vec3& position() { return _transform.translation(); }
-    glm::vec3 direction() {
-        glm::vec3 direction;
-        direction = glm::vec3(_transform.rotationMatrix() * glm::vec4(0, 0, -1, 1));
-        return direction;
-    }
+    glm::vec3 direction();
 
 private:
     Transform _transform;
