+++ include/TotoGL/RenderObject/ShaderMaterial.hpp
@@ -11,41 +11,20 @@ class ShaderMaterial {
 public:
     using ShaderType::FRAGMENT;
     using ShaderType::VERTEX;
-    ShaderMaterial()
-        : ShaderMaterial(
-            Shader<VERTEX>(),
-            Shader<FRAGMENT>()) { }
+    ShaderMaterial();
 
-    ShaderMaterial(Shader<VERTEX>&& vertex, Shader<FRAGMENT>&& frament)
-        : _vertex(std::move(vertex))
-        , _fragment(std::move(frament)) {
-        compile();
-    }
+    ShaderMaterial(Shader<VERTEX>&& vertex, Shader<FRAGMENT>&& frament);
 
     Shader<VERTEX>& vertex() { return _vertex; }
     Shader<FRAGMENT>& fragment() { return _fragment; }
 
-    void compile() {
-        _program.detachAll().attach(_vertex).attach(_fragment).link();
-    }
+    void compile();
 
-    void use() {
-        _program.use();
-    }
-    static void unuse() {
-        ShaderProgram::unuse();
-    }
+    void use();
+    static void unuse();
 
-    void uniform(const std::string& name, const UniformVariant& value) {
-        use();
-        _program.uniform(name, value);
-        unuse();
-    }
-    void uniform(const std::string& name, Texture& value) {
-        use();
-        _program.uniform(name, value);
-        unuse();
-    }
+    void uniform(const std::string& name, const UniformVariant& value);
+    void uniform(const std::string& name, Texture& value);
 
 private:
     Shader<VERTEX> _vertex;
