+++ source/RenderObject/Light.cpp
@@ -1,0 +1,29 @@
+#include "TotoGL/RenderObject/Light.hpp"
+
+#include "TotoGL/Primitives/Transform.hpp"
+#include <glm/glm.hpp>
+
+namespace TotoGL {
+
+Light::Light(glm::vec3 color, float strength, LightType type)
+    : _type(type)
+    , _color(color)
+    , _strength(strength) { }
+
+Light& Light::translate(const glm::vec3& translation) {
+    _transform.translate(translation);
+    return *this;
+}
+
+Light& Light::rotate(const float& angle, const glm::vec3& axis) {
+    _transform.rotate(angle, axis);
+    return *this;
+}
+
+glm::vec3 Light::direction() {
+    glm::vec3 direction;
+    direction = glm::vec3(_transform.rotationMatrix() * glm::vec4(0, 0, -1, 1));
+    return direction;
+}
+
+} // namespace TotoGL
